Creating a relationship between todo and the user and checking whether the user is the owner of todo or not

+++ app/Http/Controllers/TodoController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Todo;
+use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator;
 use Inertia\Inertia;
@@ -16,24 +17,32 @@ class TodoController extends Controller
 
   public function store(Request $request)
   {
-    $validator = Validator::make($request->all(), [
-      'text' => 'required|string|max:255',
-    ]);
+    $validator = Validator::make($request->all(), ['text' => 'required|string|max:255',]);
     if ($validator->fails()) return response()->json(['errors' => $validator->errors()], 422);
-
-    $todo = Todo::create($request->all());
+    $user = User::findorfail($request->userId);
+    $todo = $user->todos()->create($request->all());
     return response()->json($todo, 201);
   }
 
   public function update(Request $request, Todo $todo)
   {
-    $todo->update($request->all());
-    return response()->json($todo, 200);
+    try {
+      if ($todo->user_id !== $request->userId) {
+        throw new \Exception("You are not authorized to update this todo.");
+      }
+      $todo->update($request->all());
+      return response()->json($todo, 200);
+    } catch (\Exception $e) {
+      return response()->json(['errors' => $e->getMessage()], 422);
+    }
   }
 
-  public function destroy(Todo $todo)
+  public function destroy(Request $request, Todo $todo)
   {
     try {
+      if ($todo->user_id !== $request->userId) {
+        throw new \Exception("You are not authorized to delete this todo.");
+      }
       $todo->delete();
       return response()->json(null, 204);
     } catch (\Exception $e) {
@@ -41,9 +50,10 @@ class TodoController extends Controller
     }
   }
 
-  public function get()
+  public function get(Request $request)
   {
-    $todos = Todo::orderBy('created_at', 'desc')->get();
+    $user = User::findorfail($request->userId);
+    $todos = $user->todos()->get();
     return response()->json($todos, 200);
   }
 }
